[handlers] move handlers to separate package
